+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Fix a race condition where `req.linz.model.linz.formtools.form` would be undefined.
 - Fix Linz not being able to find labels for embedded documents.
 - Update the default for the `linz.formtools.widgets.documents` widget create button to be `Create`, and made it configurable by passing in `buttonLabel: 'New label'` to the widget.
+- Fix ckeditor for non `/admin` admin paths.
 
 ## v1.0.0-14.0.0 (14 November 2017)